requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
coloredlogs==15.0.1
    # via jule (setup.py)
greenlet==3.0.3
    # via sqlalchemy
humanfriendly==10.0
    # via coloredlogs
linkify-it-py==2.0.3
    # via markdown-it-py
markdown-it-py[linkify,plugins]==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
mdit-py-plugins==0.4.1
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
numpy==2.0.1
    # via
    #   pandas
    #   pandasql
pandas==2.2.2
    # via
    #   jule (setup.py)
    #   pandasql
pandasql==0.7.3
    # via jule (setup.py)
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.4.0
    # via python-ldap
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-ldap==3.4.4
    # via jule (setup.py)
pytz==2024.1
    # via pandas
rich==13.7.1
    # via textual
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.31
    # via pandasql
tabulate==0.9.0
    # via jule (setup.py)
textual[syntax]==0.74.0
    # via jule (setup.py)
tree-sitter==0.20.4
    # via
    #   textual
    #   tree-sitter-languages
tree-sitter-languages==1.10.2
    # via textual
typing-extensions==4.12.2
    # via
    #   sqlalchemy
    #   textual
tzdata==2024.1
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py

# The following packages are considered to be unsafe in a requirements file:
# setuptools
